requirements.txt
python-telegram-bot==11.1.0
python-dotenv==0.15.0
vk-api==11.9.1
dialogflow==1.1.0
google-cloud-dialogflow==2.0.0